fix(2.0): widen audit.domain_event.published_attempts SmallInteger → Integer
Found during the first manual end-to-end order on the all-Rust stack. The
outbox relay increments published_attempts on every delivery attempt
(bss_events::relay MARK_OK_SQL / MARK_FAIL_SQL). SmallInteger caps at 32767;
when the Rust MqChannel failed to recover from a dropped AMQP channel, the relay
retried a single unpublished row several times a second for ~3 days and
`published_attempts + 1` eventually raised "smallint out of range" — which aborts
the whole drain_once transaction and wedges the relay tick, so NO event
publishes and the async plane (order decomposition, provisioning, balance
decrements) silently dies.

Widen to Integer (migration 0031 + model). Non-destructive in-place widening.
Already applied to the live tech-vm DB. The underlying no-reconnect defect in
bss_events::mq is tracked separately.

+++ packages/bss-models/bss_models/audit.py
@@ -64,8 +64,12 @@ class DomainEvent(Base):
     # v1.2 — outbox relay bookkeeping. The relay (bss_events.relay) increments
     # published_attempts on each delivery try and records last_publish_error, so a
     # row that won't publish is visible to triage instead of silently stuck.
+    # v2.0 — Integer, not SmallInteger: a stuck relay (e.g. a broker blip the Rust
+    # MqChannel didn't recover from) can retry a single row well past SmallInteger's
+    # 32767 ceiling, and `published_attempts + 1` then raised "smallint out of range"
+    # and wedged the whole relay tick. Integer removes that failure mode.
     published_attempts: Mapped[int] = mapped_column(
-        SmallInteger, nullable=False, default=0, server_default="0"
+        Integer, nullable=False, default=0, server_default="0"
     )
     last_publish_error: Mapped[str | None] = mapped_column(Text)
